refactor: 컴포넌트 쪼개기 및 스타일 문제 수정 - 이모지 버튼 스타일, 그리드 레이아웃, 카카오 SDK 로딩 방식 개선
