+++ app/post/_components/MessageCardListWithModal.module.scss
@@ -3,7 +3,6 @@
 .message-card-list-with-modal {
     margin-top: 11.3rem;
     margin-bottom: 24.6rem;
-
     margin-left: 2.4rem;
     margin-right: 2.4rem;
     width: 100%;
