+++ components/ServiceHeader/ShareButtonMenu/ShareButtonMenu.module.scss
@@ -1,0 +1,8 @@
+.share-button-menu {
+  display: flex;
+  align-items: center;
+  
+  .popover-size-container {
+    min-width: 180px;
+  }
+}
